+++ src/com/kaylerrenslow/armaplugin/lang/sqf/syntax/ValueType.java
@@ -61,7 +61,14 @@ public enum ValueType {
 	FOR("For Type"),
 	SWITCH("Switch Type"),
 	WHILE("While Type"),
-	WITH("With Type"),;
+	WITH("With Type"),
+
+	/**
+	 * Not an actual Arma 3 data type.
+	 * This is for Arma Intellij Plugin to signify a variable is being used
+	 * and that the type is indeterminate with static type checking.
+	 */
+	_VARIABLE("`VARIABLE`");
 
 	private final String displayName;
 	private final boolean isArray;